[Feature] 헤더로 유저 정보가져오기 구현

+++ src/main/java/com/nhnacademy/bookapi/booklike/controller/BookLikeController.java
@@ -1,14 +1,11 @@
 package com.nhnacademy.bookapi.booklike.controller;
 
-import com.nhnacademy.bookapi.advice.ValidationFailedException;
-import com.nhnacademy.bookapi.booklike.domain.request.BookLikeCreateRequest;
+import com.nhnacademy.bookapi.advice.InvalidUserIdHeaderException;
 import com.nhnacademy.bookapi.booklike.domain.response.BookLikeResponse;
 import com.nhnacademy.bookapi.booklike.service.BookLikeService;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
@@ -20,9 +17,12 @@ public class BookLikeController {
 
     private final BookLikeService bookLikeService;
 
-    // 좋아요 누른 도서는 마이페이지에서 확인가능 > 임시 경로
-    @GetMapping("/users/{userId}/bookLikes")
-    public ResponseEntity<List<BookLikeResponse>> getBookLikes(@PathVariable String userId) {
+    // 마이페이지에서 좋아요 확인
+    @GetMapping("/users")
+    public ResponseEntity<List<BookLikeResponse>> getBookLikes(@RequestHeader("X-USER-ID") String userId) {
+        if (userId == null || userId.isBlank()) {
+            throw new InvalidUserIdHeaderException();
+        }
         List<BookLikeResponse> bookLikes = bookLikeService.getBookLikesByUserId(userId);
         return ResponseEntity.status(HttpStatus.OK).body(bookLikes);
     }
@@ -35,27 +35,22 @@ public class BookLikeController {
 
     @PostMapping("/books/{bookId}/bookLikes")
     public ResponseEntity<BookLikeResponse> createBookLike(@PathVariable Long bookId,
-                                                           @Valid @RequestBody BookLikeCreateRequest request,
-                                                           BindingResult bindingResult) {
-        if (bindingResult.hasErrors()) {
-            throw new ValidationFailedException();
+                                                           @RequestHeader("X-USER-ID") String userId) {
+        if (userId == null || userId.isBlank()) {
+            throw new InvalidUserIdHeaderException();
         }
-        BookLikeResponse response = bookLikeService.createBookLike(bookId, request);
+        BookLikeResponse response = bookLikeService.createBookLike(bookId, userId);
         URI location = URI.create("/books/" + bookId + "/bookLikes/" + response.bookLikeId());
         return ResponseEntity.created(location).body(response);
     }
 
-    // 마이페이지에서 삭제? 임시경로임  유저 아이디와 도서 아이디를 어떻게 받을지    헤더에 정보 저장?
-    @DeleteMapping("/users/{userId}/bookLikes/{bookId}")
-    public ResponseEntity<Void> deleteBookLikeByUserIdAndBookId(@PathVariable String userId, @PathVariable Long bookId) {
-        bookLikeService.deleteBookLikeByUserIdAndBookId(userId, bookId);
-        return ResponseEntity.noContent().build();
-    }
-
-    // 도서 아이디로 해당 도서에 대한 좋아요 전체 삭제
     @DeleteMapping("/books/{bookId}/bookLikes")
-    public ResponseEntity<Void> deleteBookLikesByBookId(@PathVariable Long bookId) {
-        bookLikeService.deleteBookLikeByBookId(bookId);
+    public ResponseEntity<Void> deleteBookLikeByUserIdAndBookId(@RequestHeader("X-USER-ID") String userId,
+                                                                @PathVariable Long bookId) {
+        if (userId == null || userId.isBlank()) {
+            throw new InvalidUserIdHeaderException();
+        }
+        bookLikeService.deleteBookLikeByUserIdAndBookId(userId, bookId);
         return ResponseEntity.noContent().build();
     }
 }
